requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==25.1.0
Automat==24.8.1
autopep8==2.3.2
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==44.0.1
css-html-js-minify==2.5.5
cssselect2==0.8.0
Django==5.2.2
django-cors-headers==4.7.0
django-crispy-forms==2.0
django-debug-toolbar==5.2.0
django-extensions==3.2.3
django-formtools==2.5.1
django-otp==1.6.0
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
flake8==7.2.0
gunicorn==23.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
lxml==5.4.0
mccabe==0.7.0
msgpack==1.1.0
mysqlclient==2.1.0
oscrypto==1.3.0
packaging==24.2
phonenumbers==9.0.2
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pyOpenSSL==25.0.0
pyotp==2.9.0
pypdf==5.6.1
pypng==0.20220715.0
python-bidi==0.6.6
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qrcode==7.4.2
redis==5.2.1
reportlab==4.4.2
requests==2.32.4
service-identity==24.2.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tomli==2.2.1
typing_extensions==4.12.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.3.0
xhtml2pdf==0.2.17
zope.interface==7.2
python-docx==1.1.0